style: center navbar links and improve navbar layout
- Use CSS grid for navbar layout
- Center nav links in the middle of the page
- Ensure navbar spacer logic is robust and works with not-found pages
