+++ components/layout/navbar/navbar.module.css
@@ -39,6 +39,7 @@
   gap: var(--space-sm);
   cursor: pointer;
   text-decoration: none;
+  justify-self: start;
 }
 
 .navbarTitle {
@@ -107,14 +108,23 @@
 }
 .wrapper {
   width: 100%;
-  display: flex;
-  justify-content: space-between;
+  display: grid;
+  grid-template-columns: 1fr auto 1fr;
+  align-items: center;
   position: relative;
 }
 .navbarDesktop {
   display: flex;
   align-items: center;
   gap: var(--space-md);
+}
+
+.navbarDesktop:first-of-type {
+  justify-self: center;
+}
+
+.navbarDesktop:nth-of-type(2) {
+  justify-self: end;
 }
 .navbarMobile {
   display: none;
